+++ src/main/java/me/earth/earthhack/impl/modules/client/management/ListenerGameLoop.java
@@ -1,3 +1,6 @@
+/*
+ * Decompiled with CFR 0.150.
+ */
 package me.earth.earthhack.impl.modules.client.management;
 
 import me.earth.earthhack.impl.event.events.misc.GameLoopEvent;
@@ -13,7 +16,8 @@ extends ModuleListener<Management, GameLoopEvent> {
     @Override
     public void invoke(GameLoopEvent event) {
         if (ListenerGameLoop.mc.world != null && ((Management)this.module).time.getValue() != 0) {
-            ListenerGameLoop.mc.world.setWorldTime(-((Management)this.module).time.getValue().intValue());
+            ListenerGameLoop.mc.world.setWorldTime((long)(-((Management)this.module).time.getValue().intValue()));
         }
     }
 }
+
