+++ src/main/java/me/earth/earthhack/impl/util/minecraft/KeyBoardUtil.java
@@ -1,3 +1,11 @@
+/*
+ * Decompiled with CFR 0.150.
+ * 
+ * Could not load the following classes:
+ *  net.minecraft.client.settings.KeyBinding
+ *  org.lwjgl.input.Keyboard
+ *  org.lwjgl.input.Mouse
+ */
 package me.earth.earthhack.impl.util.minecraft;
 
 import me.earth.earthhack.api.setting.Setting;
@@ -20,6 +28,7 @@ public class KeyBoardUtil {
     }
 
     public static boolean isKeyDown(int key) {
-        return key != 0 && key != -1 && (key < 0 ? Mouse.isButtonDown(key + 100) : Keyboard.isKeyDown(key));
+        return key != 0 && key != -1 && (key < 0 ? Mouse.isButtonDown((int)(key + 100)) : Keyboard.isKeyDown((int)key));
     }
 }
+
